Use typescript
- Moved comments to the index.ts.
- Genereates index.js/index.d.ts from the index.ts.
- Generated index.d.ts is similar to the original d.ts.
- Generated index.js is little different, but the tests pass.
