+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
 yarn.lock
+index.js
+index.d.ts
